fix(app): flickering when queue is updated
The app was flicking because of a stateful array
being updated. This triggered a full component
update which had everything flash.

This change minimizes the scope of the queue to
a new queue container element. That way only it
will exhibit the flicker if present.

To achieve this in a more clear way to be immutable
immer was brought in as suggested by the lit team
for these situations.